+++ src/main/java/org/cotato/csquiz/domain/attendance/poi/AttendanceRecordExcelData.java
@@ -1,4 +1,4 @@
-package org.cotato.csquiz.common.poi;
+package org.cotato.csquiz.domain.attendance.poi;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,6 +6,9 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.cotato.csquiz.common.poi.CellData;
+import org.cotato.csquiz.common.poi.ExcelColumnName;
+import org.cotato.csquiz.common.poi.ExcelData;
 import org.cotato.csquiz.domain.attendance.entity.AttendanceRecord;
 import org.cotato.csquiz.domain.attendance.enums.AttendanceResult;
 import org.cotato.csquiz.domain.attendance.vo.AttendRecordStatics;
