+++ src/main/java/org/cotato/csquiz/domain/attendance/service/AttendanceExcelService.java
@@ -12,8 +12,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.cotato.csquiz.api.attendance.dto.AttendanceStatistic;
 import org.cotato.csquiz.api.attendance.dto.GenerationMemberAttendanceRecordResponse;
-import org.cotato.csquiz.common.poi.AttendanceRecordExcelData;
-import org.cotato.csquiz.common.poi.AttendanceRecordExcelData.AttendRecord;
+import org.cotato.csquiz.domain.attendance.poi.AttendanceRecordExcelData;
+import org.cotato.csquiz.domain.attendance.poi.AttendanceRecordExcelData.AttendRecord;
 import org.cotato.csquiz.common.poi.ExcelWriter;
 import org.cotato.csquiz.domain.attendance.entity.Attendance;
 import org.cotato.csquiz.domain.attendance.entity.AttendanceRecord;
